refactor: make async/sync functions DRY
